Detect more magic constants
