--- src/Rules/Constants/MagicClassConstantRule.php
@@ -1,33 +1,0 @@
-<?php declare(strict_types = 1);
-
-namespace PHPStan\Rules\Constants;
-
-use PhpParser\Node;
-use PHPStan\Analyser\Scope;
-use PHPStan\Rules\Rule;
-use PHPStan\Rules\RuleErrorBuilder;
-use function sprintf;
-
-/** @implements Rule<Node\Scalar\MagicConst\Class_> */
-class MagicClassConstantRule implements Rule
-{
-
-	public function getNodeType(): string
-	{
-		return Node\Scalar\MagicConst\Class_::class;
-	}
-
-	public function processNode(Node $node, Scope $scope): array
-	{
-		if ($scope->isInClass()) {
-			return [];
-		}
-
-		return [
-			RuleErrorBuilder::message(
-				sprintf('Magic constant %s is always empty when used outside a class.', $node->getName()),
-			)->build(),
-		];
-	}
-
-}
